changed return value of getDigitalRoot to int since the maximum returned value will be 9

+++ src/test/java/net/emteeware/common/DigitalRootNegativeTest.java
@@ -10,6 +10,6 @@ public class DigitalRootNegativeTest {
 
     @Test(expected=IllegalArgumentException.class)
     public void testGetDigitalRootWithNegativeInput() throws Exception {
-        long result = DigitSum.getDigitalRoot(-1);
+        int result = DigitSum.getDigitalRoot(-1);
     }
 }
